secmet: move frameshift function to secmet.locations
