+++ antismash/common/secmet/test/test_locations.py
@@ -11,6 +11,7 @@ from unittest.mock import patch
 
 from antismash.common import secmet
 from antismash.common.secmet.locations import (
+    _adjust_location_by_offset as adjust,
     convert_protein_position_to_dna,
     build_location_from_others,
     ensure_valid_locations,
@@ -529,3 +530,51 @@ class TestEnsureValid(unittest.TestCase):
                     SeqFeature(build_compound([(10, 30), (0, 9)], -1), type="CDS")]
         with self.assertRaisesRegex(ValueError, "inconsistent exon ordering"):
             self.check(features, circular=True)
+
+
+class TestLocationAdjustment(unittest.TestCase):
+    def setUp(self):
+        # not all of these really make sense biologically, but they're all valid computationally
+        self.position_types = [ExactPosition, AfterPosition, BeforePosition, int]
+
+    def test_single_forward(self):
+        for position_type in self.position_types:
+            old = FeatureLocation(position_type(5), 12, 1)
+            for offset in range(-2, 3):
+                new = adjust(old, offset)
+                assert isinstance(new.start, position_type)
+                assert new.start == old.start + offset
+                assert new.end is old.end
+
+    def test_single_reverse(self):
+        for position_type in self.position_types:
+            old = FeatureLocation(5, position_type(12), -1)
+            for offset in range(-2, 3):
+                new = adjust(old, offset)
+                assert isinstance(new.end, position_type)
+                assert new.end == old.end + offset
+                assert new.start is old.start
+
+    def test_compound_forward(self):
+        for position_type in self.position_types:
+            old = CompoundLocation([FeatureLocation(position_type(5), 12, 1),
+                                    FeatureLocation(15, 17, 1)])
+            for offset in range(-2, 3):
+                new = adjust(old, offset)
+                assert isinstance(new.start, position_type)
+                assert new.start == old.start + offset
+                assert new.parts[0].end is old.parts[0].end
+                for old_part, new_part in zip(old.parts[1:], new.parts[1:]):
+                    assert old_part is new_part
+
+    def test_compound_reverse(self):
+        for position_type in self.position_types:
+            old = CompoundLocation([FeatureLocation(15, position_type(17), -1),
+                                    FeatureLocation(5, 12, -1)])
+            for offset in range(-2, 3):
+                new = adjust(old, offset)
+                assert isinstance(new.end, position_type)
+                assert new.end == old.end + offset
+                assert new.parts[0].start is old.parts[0].start
+                for old_part, new_part in zip(old.parts[1:], new.parts[1:]):
+                    assert old_part is new_part
